chore: move more pages to app
